Add files via upload
EXPLANATION OF CODE IF NEEDED:
This is the first code that I wrote using Java , the source is apnikaksha
Link :  https://youtu.be/yRpLlJmRo2w?feature=shared (in case somebody need)
Breaking down the keywords
1)Public class main{}
Public: It is basically an access modifier that makes the class accessible from anywhere
Class: It is a keyword that tells Java that you are now defining an class
Main: This is the name of the class. Here my fine was saved as Main.java therefore the name of the class of Main,basically your class name must match your file name
{}: This is the class body, basically where you write down all your codes  .

Java needs a class to run all its code in and this is that place here .

2)public static void main(String[] args){}   
[NOTE: Java is a case sensitive language so the 's' in Strings should be capital, the 'm' in main should be small also it is not the same as your file name]
(when you click run java looks for this method in the class you named)
public: Again the method is accessible everywhere
static: It means the method belongs to a class only not to any object and used so java can run main() without using any object
void: It does not return anything to the system ,it just starts and run your codes
main:This is the method name java looks for to start running the program 
String[] args: This means that java can take inputs from terminal while running the program

 3)//variables: The sign "//" is used to comment out things in java so I commented the term variables showcasing that I am taking variables now

4)int a=108043;
   int b=250333;  :   int refers to integers "int a" means I am taking a variable as integer and then assigned it a value, also it's important to end these lines with ";" sign otherwise error occur.

5)//int sum=a+b;
   //int diff=b-a;
  int multiply=a*b;  : previous 2 commands I used them so commented out the third one is multiplication where I am multiplying the assigned variable with "*" operator used for multiplication , the variables should not be double quotes .

6) System.out.println(multiply);
system: A built in class in java 
out: java's way of saying I want to show some output
println(): Its a method comman you call and it prints anything you give 


then you just close your curly brackets and enjoy running your code, have a great day!